Unwrap stored fields readers to ensure that per format optimizations work
